Reorganize test fixtures into focused modules with module docstrings.
