+++ featherstore/_table/append.py
@@ -14,7 +14,7 @@ def can_append_table(table, df, warnings):
 
     table_data = table._table_data
     index_name = table_data["index_name"]
-    cols = _table_utils.get_data_col_names(df, index_name=index_name)
+    cols = _table_utils.get_col_names(df, index_name=index_name)
     _raise_if.incoming_index_schema_incompatible_with_stored(
         df, table_data, cols, check_index_values=False
     )
