+++ featherstore/_table/insert_cols.py
@@ -14,7 +14,7 @@ def can_insert_columns(table, df, idx, warnings):
 
     table_data = table._table_data
     index_name = table_data["index_name"]
-    cols = _table_utils.get_data_col_names(df, index_name=index_name)
+    cols = _table_utils.get_col_names(df, index_name=index_name)
     index = _table_utils.get_index_if_exists(df, index_name)
 
     _raise_if.incoming_index_schema_incompatible_with_stored(
@@ -100,4 +100,3 @@ def create_partitions(df, rows_per_partition, partition_names):
     return _partitions.create_partitions(
         df, rows_per_partition, partition_names, strategy="grow"
     )
-
